docs(site): remove admin-ui from docs site
