+++ packages/next-docs/components/icons-grid/icons-grid.module.css
@@ -1,0 +1,4 @@
+.iconsGrid {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+}
