--- packages/next-docs/components/icons-grid/icons.css.ts
@@ -1,32 +1,0 @@
-import { csx } from '@vtex/admin-ui'
-
-export const iconsGridStyle = csx({
-  display: 'grid',
-  gridTemplateColumns: 'repeat(3, 1fr)',
-})
-
-export const iconsCardStyle = csx({
-  display: 'flex',
-  paddingY: '1.5rem',
-  paddingX: '1rem',
-  flexDirection: 'column',
-  justifyContent: 'flex-start',
-  alignItems: 'center',
-  svg: {
-    marginBottom: '1rem',
-  },
-})
-
-export const iconDescriptionStyle = csx({
-  fontWeight: 400,
-  fontSize: '0.75em',
-  textAlign: 'center',
-  color: '#6C727F',
-})
-
-export const iconNameStyle = csx({
-  fontWeight: 500,
-  textAlign: 'center',
-  fontSize: '0.8em',
-  fontFamily: 'monospace',
-})
