+++ packages/next-docs/components/tokens-grid/cell.tsx
@@ -1,11 +1,11 @@
 import type { ComponentPropsWithoutRef } from 'react'
 import React from 'react'
-import { cx } from '@vtex/admin-ui'
+import { cx } from '@vtex/shoreline-utils'
 
-import { cellTheme } from './tokens-grid.css'
+import styles from './cell.module.css'
 
 export function Cell(props: ComponentPropsWithoutRef<'div'>) {
   const { className = '', ...restProps } = props
 
-  return <div className={cx(cellTheme, className)} {...restProps} />
+  return <div className={cx(styles.cell, className)} {...restProps} />
 }
