+++ packages/next-docs/components/tokens-grid/grid.module.css
@@ -1,0 +1,20 @@
+.grid {
+  display: grid;
+  grid-template-columns: 2fr 2fr 2fr 5rem;
+}
+
+.grid[data-foundation='breakpoint'] {
+  grid-template-columns: 2fr 2fr 2fr;
+}
+
+.grid[data-foundation='elevation'] {
+  grid-template-columns: 2fr 2fr 3fr 5rem;
+}
+
+.grid[data-foundation='space'] {
+  grid-template-columns: 2fr 2fr 1fr 1fr;
+}
+
+.grid[data-foundation='border-radius'] {
+  grid-template-columns: 2fr 2fr 1fr 5rem;
+}
