+++ packages/next-docs/components/tokens-grid/head.tsx
@@ -1,7 +1,7 @@
 import type { ComponentPropsWithoutRef } from 'react'
 import React from 'react'
-import { headTheme } from './tokens-grid.css'
+import styles from './head.module.css'
 
 export function Head(props: ComponentPropsWithoutRef<'div'>) {
-  return <div className={headTheme} {...props} />
+  return <div className={styles.head} {...props} />
 }
