--- packages/next-docs/examples/playground.ts
@@ -1,9 +1,0 @@
-export const overview = {
-  '/App.js': `import { Button } from '@vtex/admin-ui'
-
-export default function App() {
-  return (
-    <Button>Admin button</Button>
-  )
-}`,
-}
