--- packages/next-docs/examples/table/anatomy.ts
@@ -1,131 +1,0 @@
-export const files = {
-  '/App.js': `import { 
-  createColumns,
-  Table,
-  THead,
-  THeadCell,
-  TBody,
-  TBodyRow,
-  TBodyCell,
-  useTableState,
-  DataView,
-  useDataViewState,
-  IconEye,
-  csx,
-  Tag,
-} from '@vtex/admin-ui'
-const columns = createColumns([
-  {
-    id: 'name',
-    header: 'Name',
-    width: '2fr',
-    resolver: {
-      type: 'text',
-      columnType: 'name',
-      mapText: (item) => item.name,
-      render: ({ data }) => (
-        <div className={csx({ minWidth: '10rem' })}>{data}</div>
-      ),
-    },
-  },
-  {
-    id: 'brand',
-    header: 'Brand',
-    width: '1fr',
-    resolver: {
-      type: 'root',
-      render: ({ item }) => {
-        return <div className={csx({ minWidth: '6rem' })}>{item.brand}</div>
-      },
-    },
-  },
-  {
-    id: 'price',
-    header: 'Price',
-    width: '1fr',
-    resolver: {
-      type: 'currency',
-      locale: 'en-US',
-      currency: 'USD',
-    },
-  },
-  {
-    id: 'status',
-    header: 'Status',
-    width: '1fr',
-    resolver: {
-      type: 'root',
-      render: ({ item }) => {
-        return <Tag label={item.status} size="normal" />
-      },
-    },
-  },
-  {
-    resolver: {
-      type: 'menu',
-      actions: [
-        {
-          label: 'View details',
-          icon: <IconEye />,
-        },
-      ],
-    },
-  },
-])
-const items = [
-  {
-    id: 495953,
-    name: 'Wooden Eyeglasses',
-    status: 'Inactive',
-    brand: 'Revolution',
-    price: 42.14,
-  },
-  {
-    id: 429048,
-    name: 'Brazil Straw Hat',
-    status: 'Inactive',
-    brand: 'Pathway',
-    price: 38.52,
-  },
-  {
-    id: 495954,
-    name: 'Striped Beach Short',
-    status: 'Inactive',
-    brand: 'Desire Spirit',
-    price: 23.49,
-  },
-]
-export default function App() {
-  const view = useDataViewState()
-  const { data, getBodyCell, getHeadCell, getTable } = useTableState({
-    status: view.status,
-    columns,
-    items,
-  })
-  return (
-    <DataView state={view}>
-      <Table {...getTable()}>
-        <THead>
-          {columns.map((column) => (
-            <THeadCell {...getHeadCell(column)} />
-          ))}
-        </THead>
-        <TBody>
-          {data.map((item) => {
-            return (
-              <TBodyRow
-                key={item.id}
-                onClick={() => alert('Item: item.name')}
-              >
-                {columns.map((column) => {
-                  return <TBodyCell {...getBodyCell(column, item)} />
-                })}
-              </TBodyRow>
-            )
-          })}
-        </TBody>
-      </Table>
-    </DataView>
-  )
-}`,
-}
