+++ packages/next-docs/package.json
@@ -11,7 +11,6 @@
     "@types/he": "1.2.3"
   },
   "dependencies": {
-    "@vtex/admin-ui": "^0.136.1",
     "@vtex/shoreline-components": "workspace:*",
     "@vtex/shoreline-date": "workspace:*",
     "@vtex/shoreline-icons": "workspace:*",
